refactor: `Avatar` component 경로 이동
